remove: migrate FinancialResultWatch to `usa_research` app and delete related templates from `vietnam_research`
