+++ vietnam_research/templates/vietnam_research/index.html
@@ -22,7 +22,6 @@
                         <a class="btn btn-outline-primary" href="{% url 'vnm:economy' %}">経済指標</a>
                         <a class="btn btn-outline-primary" href="{% url 'vnm:tools' %}">株式ツール</a>
                         <a class="btn btn-outline-primary" href="{% url 'vnm:watchlist' %}">ウォッチリスト</a>
-                        <a class="btn btn-outline-secondary" href="{% url 'vnm:financial_results' %}">決算情報</a>
                     </div>
                 </div>
             </div>
